# Simpler and faster implementation of Floyd's F2