@ciphergoth 2022-12-29
On my machine - an M1 MacBook I get the following benchmarks:

before:
```
test misc_sample_indices_100_of_1G           ... bench:       2,672 ns/iter (+/- 37)
test misc_sample_indices_100_of_1M           ... bench:       2,649 ns/iter (+/- 64)
test misc_sample_indices_100_of_1k           ... bench:         454 ns/iter (+/- 10)
test misc_sample_indices_10_of_1k            ... bench:          50 ns/iter (+/- 1)
test misc_sample_indices_1_of_1k             ... bench:          20 ns/iter (+/- 0)
test misc_sample_indices_200_of_1G           ... bench:       4,078 ns/iter (+/- 160)
test misc_sample_indices_400_of_1G           ... bench:       8,120 ns/iter (+/- 139)
test misc_sample_indices_600_of_1G           ... bench:      11,577 ns/iter (+/- 186)
```

After:
```
test misc_sample_indices_100_of_1G           ... bench:       2,252 ns/iter (+/- 16)
test misc_sample_indices_100_of_1M           ... bench:       2,233 ns/iter (+/- 49)
test misc_sample_indices_100_of_1k           ... bench:         453 ns/iter (+/- 3)
test misc_sample_indices_10_of_1k            ... bench:          54 ns/iter (+/- 0)
test misc_sample_indices_1_of_1k             ... bench:          19 ns/iter (+/- 0)
test misc_sample_indices_200_of_1G           ... bench:       4,060 ns/iter (+/- 581)
test misc_sample_indices_400_of_1G           ... bench:       8,080 ns/iter (+/- 618)
test misc_sample_indices_600_of_1G           ... bench:      11,537 ns/iter (+/- 3,903)```