@ciphergoth 2022-12-29
The 8% slower performance on `misc_sample_indices_10_of_1k` seemed bad, so I have pushed a new version with no performance regressions.
```
test misc_sample_indices_100_of_1G           ... bench:       2,160 ns/iter (+/- 52)
test misc_sample_indices_100_of_1M           ... bench:       2,154 ns/iter (+/- 57)
test misc_sample_indices_100_of_1k           ... bench:         453 ns/iter (+/- 6)
test misc_sample_indices_10_of_1k            ... bench:          49 ns/iter (+/- 3)
test misc_sample_indices_1_of_1k             ... bench:          19 ns/iter (+/- 0)
test misc_sample_indices_200_of_1G           ... bench:       4,068 ns/iter (+/- 29)
test misc_sample_indices_400_of_1G           ... bench:       8,081 ns/iter (+/- 1,526)
test misc_sample_indices_600_of_1G           ... bench:      11,534 ns/iter (+/- 310)
```